refactor(options): simplify Options component and update navigation logic
- Remove unused imports and components to streamline the Options component
- Update useEffect to navigate to the dashboard on component mount
- Refactor shadow DOM functions for better clarity and maintainability
